stride: split CLI commands into per-command files
Program.cs becomes a thin composition root; each command (sdk, upgrade, studio+asset
in ToolCommands) moves to its own file mirroring NewCommand, with per-command options
so each --version/--prerelease carries help tailored to that command. Shared version
resolution moves to CliVersion.ResolveOrReport.
